First attempt to add remaining ingredients and amounts of ingredients to the plug-in (and likely not the last)

+++ editor/src/main/java/nl/knokko/customitems/editor/set/recipe/ingredient/Ingredient.java
@@ -94,7 +94,7 @@ public abstract class Ingredient implements SCIngredient {
 		saveSpecifics(output);
 	}
 
-	public int getAmount() {
+	public byte getAmount() {
 		return amount;
 	}
